// src/types/FormTypes.ts
export type FormDataTypes = {
  firstName: string;
  lastName: string;
  passportNumber: string;
  nationalID: string;
  dateOfBirth: string;
  gender: string;
  email: string;
  phone: string;
  address: string;
  zipCode: string;
  effectiveDate: string;
  duration: number | 0;
//   spouses: SpouseType[];
//   children: ChildType[];
//   vehicles: VehicleType[];
};